New command sending infrastructure
Implemented the _raw_command method which largely works around imaplib
for sending commands. It will automatically use literals when 8-bit
data must be sent and supports abitrary numbers of literals (imaplib
only supports sending a single literal and it's awkward to do
so). This is especially critical for correctly allowing complex search
queries that may involve multiple 8-bit criteria.

The search, sort, thread and gmail_search methods now all use the new
infrastructure. Soon, some version of _raw_command will be become the
only way that IMAPClient sends commands. This work also paves the way
for removing the dependency on imaplib.

This work is not yet complete. A number of test and documentation
updates are still required.