Use the standard logging module
Instead of a custom way of managing log lines, this commit
leverages the standard logging module. As IMAPClient is a library,
the only thing to do is to create a logger and log away.

The final user of IMAPClient is now in charge of formatting,
writing to files or discarding log lines.