Prevent IMAP passwords from going to the logging system
Even if the final user did not enable DEBUG logging for its
application, some logs records can end up in tools like Sentry
via breadcrumbs.

This commit makes sure that secrets are never sent to the
logging system by verifying and redacting the log message before
passing it to the logger.

In the future, once we get rid of imaplib, the code replacing it
could avoid logging raw data containing sensitive information in
the first place.