Locate special folders like Sent or Trash
Configuring special folders is often a pain that requires the
owner of an account to give the location of folders himself.

With the help of IMAP extensions and a few assumptions we can try
to guess where these folders are. In practice this is enough for
the vast majority of accounts.

Related to #296