Allow the connected socket to be polled for data
This is done by exposing the underlying `socket()` used by `IMAPClient`.
The socket was previously avaiable only via the `_sock` property,
which was not and is not a supported way of accessing it.

However, since `_sock` was the only reliable way to do so,
third-party packages might be relying on it.
For transition purposes, using it will log a deprecation warning.
(Obligatory Workflow xkcd <https://xkcd.com/1172/>.)